template/requirements.txt
# Discord Bot Template - 依存関係

# Discord.py - Discord Bot開発の基本ライブラリ
discord.py>=2.3.0

# python-dotenv - 環境変数管理
python-dotenv>=0.19.0

# 以下は機能追加時に必要になる場合があります（コメントアウト）

# OpenAI API - AI機能追加時
# openai>=1.0.0

# Requests - HTTP通信が必要な場合
# requests>=2.28.0

# aiohttp - 非同期HTTP通信が必要な場合  
# aiohttp>=3.8.0

# Pillow - 画像処理が必要な場合
# Pillow>=9.0.0

# beautifulsoup4 - Webスクレイピングが必要な場合
# beautifulsoup4>=4.11.0

# youtube-dl - 音声・動画処理が必要な場合
# yt-dlp>=2023.1.0

# pydub - 音声ファイル処理が必要な場合
# pydub>=0.25.0